Move validation funcs to validation.ex / tests. Parser should now just parse
